Move lodash-fp-types type tests to core-utils
