+++ package.json
@@ -73,7 +73,8 @@
     "preinstall": "node .hooks/check-engine-light.js",
     "setenv": "echo VITE_APP_VERSION=$(git rev-parse HEAD) VITE_APP_BUILD_TIMESTAMP=$(date -u \"+%s000\")",
     "test": "jest",
-    "check-types": "check-dts types/**/*.ts",
+    "check-dts": "check-dts types/**/*.ts",
+    "check-types": "yarn workspaces foreach run check-dts",
     "serve": "vite preview",
     "start": "concurrently \"tsc --build --watch\" \"env $(yarn setenv) vite\"",
     "postinstall": "husky install"
@@ -94,7 +95,7 @@
     "@typescript-eslint/eslint-plugin": "^5.59.1",
     "@vitejs/plugin-react": "^4.0.0",
     "browserslist": "^4.18.1",
-    "check-dts": "^0.7.1",
+    "check-dts": "^0.7.2",
     "concurrently": "^8.2.0",
     "eslint": "^8.3.0",
     "eslint-config-airbnb-typescript-prettier": "^5.0.0",
