+++ packages/core-utils/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "scripts": {
     "build": "tsc && tsc -p tsconfig.cjs.json",
+    "check-dts": "check-dts",
     "test": "jest"
   },
   "type": "module",
@@ -26,6 +27,7 @@
     "@terra-ui-packages/test-utils": "*",
     "@types/jest": "^28.1.8",
     "@types/lodash": "^4.14.184",
+    "check-dts": "^0.7.2",
     "jest": "^27.4.3",
     "typescript": "^4.9.5"
   }
