+++ packages/core-utils/src/type-utils/lodash-fp-types.errors.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import _ from 'lodash/fp';
 
-import { SafeCurry2, SafeCurry3, SafeCurry4 } from '../../src/libs/type-utils/lodash-fp-types';
+import { SafeCurry2, SafeCurry3, SafeCurry4 } from './lodash-fp-types';
 
 // Curry2 ////////////////////////////////////////////////////////////////
